Give every Go test file one subject
routes_test.go held the tests for six subjects at once, so the tests for a
handler lived nowhere near it and the file had no source counterpart. Each
suite now sits beside the code it covers and is named after what it exercises,
and the shared request harness has a file of its own.
